renamed to CryptConverter, moved to crypto-core
