+++ core/common/token-core/build.gradle.kts
@@ -18,8 +18,10 @@ plugins {
 }
 
 dependencies {
+    api(project(":spi:common:token-spi"))
     api(project(":spi:common:jwt-spi"))
 
+    implementation(project(":extensions:common:crypto:crypto-core")) // for the CryptoConverter
     implementation(libs.nimbus.jwt)
     api(libs.bouncyCastle.bcpkixJdk18on)
 }
